added attributes
added UserPermissionAttribute
added BotPermissionAttribute
added ContextAttribute
added Humanizer
made a partial Help module

+++ Rias.Core/TypeParsers/IGuildUserTypeParser.cs
@@ -24,7 +24,7 @@ namespace Rias.Core.TypeParsers
                     return TypeParserResult<IGuildUser>.Successful(user);
             }
             
-            if (uint.TryParse(value, out var id))
+            if (ulong.TryParse(value, out var id))
             {
                 user = context.Guild.GetUser(id);
                 if (user != null)
@@ -39,8 +39,7 @@ namespace Rias.Core.TypeParsers
                 var username = value.Substring(0, index);
                 if (ushort.TryParse(value.Substring(index + 1), out var discriminator))
                 {
-                    user = users.FirstOrDefault(u => u.DiscriminatorValue == discriminator
-                                                                       && string.Equals(u.Username, username, StringComparison.OrdinalIgnoreCase));
+                    user = users.FirstOrDefault(u => u.DiscriminatorValue == discriminator && string.Equals(u.Username, username, StringComparison.OrdinalIgnoreCase));
                     if (user != null)
                         return TypeParserResult<IGuildUser>.Successful(user);
                 }
